inventory_backend/requirements/base.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
psycopg2-binary==2.8.3
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
cryptography==2.6.1
python-dotenv==0.10.2
django-cors-headers==3.0.1
django-health-check==3.9.0
psutil==5.6.0
drfdocs==0.0.11
requests==2.20.0
coreapi==2.3.3
Pillow==7.1.2
